PowerFlowData

PowerFlowData.jl provides a parser for PSS/E-format .raw Power Flow Data Files.

To read a .raw file, use parse_network:

parse_network("file.raw")

This will return a Network object, which contains the data parsed into dedicated structures matching the PSS/E-format specification.

Example

Usually your data will be in a file, and you'd read it with parse_network("file.raw"), but here we'll pass in the data directly, to show how it matches to the output:

julia> using PowerFlowData, DataFrames
julia> data = IOBuffer("""
           0,   100.00          / PSS/E-29.3    WED, SEP 15 2021  21:04
           SE SNAPSHOT 15-09-2021 PEAK CASE 18:00
           FULL COPY OF SYNTHETIC
                1,'AAA    3    ', 111.0000,4,     0.000,     0.000, 327,   1,0.00000,   0.0000,   1
           222222,'PRPR C D    ',  42.0000,1,     0.000,     0.000, 694,  24,1.11117,  20.0606,   7
           0 / END OF BUS DATA, BEGIN LOAD DATA
           """
       );
julia> network = parse_network(data);
julia> NamedTuple(network.caseid)  # Case Identification data is a single row.(ic = 0, sbase = 100.0)
julia> DataFrame(network.buses)    # Bus data, and all other data, is a table.2×11 DataFrame
 Row │ i       name          basekv   ide    gl       bl       area   zone   v ⋯
     │ Int64   String15…     Float64  Int64  Float64  Float64  Int64  Int64  F ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │      1  AAA    3        111.0      4      0.0      0.0    327      1  0 ⋯
   2 │ 222222  PRPR C D         42.0      1      0.0      0.0    694     24  1
                                                               3 columns omitted

API

PowerFlowData.parse_network — Function

parse_network(source) -> Network

Read a PSS/E-format .raw Power Flow Data file and return a Network object.

PowerFlowData.Network — Type

Network

Representation of a power network.

The PSS/E data format comprises 16 data categories of network and equipment elements, each of which requires a particular type of data.

Similarly, a Network stores the data from each category in its own dedicated structure.

Currently supported are:

CaseID

Buses

Loads

Generators

Branches

TwoWindingTransformers

CaseID data is a single row (in the Tables.jl-sense). You can access it like network.caseid and interact with it like a NamedTuple, or even convert it to a NamedTuple with NamedTuple(caseid).

All other records (buses, loads, etc.) can be accessed also via the fields, for example network.buses, and each is returned as lightweight table structure (again, in the Tables.jl-sense). That is, all structures implement the Tables.jl interface, so can be passed to any valid sink, such as a DataFrame like DataFrame(network.buses).

For more info on working with tables see Tables.jl, and for common table operations see TableOperations.jl.

Fields

caseid::CaseID

Case identification data.

buses::Buses

Bus records.

loads::Loads

Load records.

generators::Generators

Generator records.

branches::Branches

Non-transformer Branch records.

two_winding_transformers::TwoWindingTransformers

Two-winding Transformer records.

PowerFlowData.CaseID — Type

struct CaseID <: Tables.AbstractRow

Fields

ic::Int64

IC Change code: 0 - base case (i.e., clear the working case before adding data to it). 1 - add data to the working case.

sbase::Float64

System base MVA.

PowerFlowData.Buses — Type

struct Buses <: PowerFlowData.Records

Each network bus to be represented in PSS/E is introduced through a bus data record. Each bus data record includes not only data for the basic bus properties but also includes information on an optionally connected shunt admittance to ground. That admittance can represent a shunt capacitor or a shunt reactor (both with or without a real component) or a shunt resistor. It must not represent line connected admittance, loads, line charging or transformer magnetizing impedance, all of which are entered in other data categories.

Fields

i::Vector{Int64}

Bus number (1 to 999997).

name::Vector{WeakRefStrings.String15}

Alphanumeric identifier assigned to bus "I". The name may be up to twelve characters and must be enclosed in single quotes. NAME may contain any combination of blanks, uppercase letters, numbers and special characters, but the first character must not be a minus sign.

basekv::Vector{Float64}

Bus base voltage; entered in kV.

ide::Vector{Int64}

Bus type code: 1 - load bus or other bus without any generator boundary condition. 2 - generator or plant bus either regulating voltage or with a fixed reactive power (Mvar). A generator that reaches its reactive power limit will no longer control voltage but rather hold reactive power at its limit. 3 - swing bus or slack bus. It has no power or reactive limits and regulates voltage at a fixed reference angle. 4 - disconnected or isolated bus.

gl::Vector{Float64}

Active component of shunt admittance to ground; entered in MW at one per unit voltage. GL should not include any resistive admittance load, which is entered as part of load data.

bl::Vector{Float64}

Reactive component of shunt admittance to ground; entered in Mvar at one per unit voltage. BL should not include any reactive impedance load, which is entered as part of load data; line charging and line connected shunts, which are entered as part of non-transformer branch data; or transformer magnetizing admittance, which is entered as part of transformer data. BL is positive for a capacitor, and negative for a reactor or an inductive load.

area::Vector{Int64}

Area number. 1 through the maximum number of areas at the current size level.

zone::Vector{Int64}

Zone number. 1 through the maximum number of zones at the current size level.

vm::Vector{Float64}

Bus voltage magnitude; entered in pu.

va::Vector{Float64}

Bus voltage phase angle; entered in degrees.

owner::Vector{Int64}

Owner number. 1 through the maximum number of owners at the current size level.

PowerFlowData.Loads — Type

struct Loads <: PowerFlowData.Records

Each network bus at which a load is to be represented must be specified in at least one load data record. If multiple loads are to be represented at a bus, they must be individually identified in a load data record for the bus with a different load identifier. Each load at a bus can be a mixture of loads with different characteristics.

Fields

i::Vector{Int64}

Buses number, or extended buses name enclosed in single quotes.

id::Vector{WeakRefStrings.String3}

One- or two-character uppercase non blank alphanumeric load identifier used to distinguish among multiple loads at bus "I". It is recommended that, at buses for which a single load is present, the load be designated as having the load identifier '1'.

status::Vector{Bool}

Initial load status of one for in-service and zero for out-of-service.

area::Vector{Int64}

Area to which the load is assigned (1 through the maximum number of areas at the current size level).

zone::Vector{Float64}

Zone to which the load is assigned (1 through the maximum number of zones at the current size level).

pl::Vector{Float64}

Active power component of constant MVA load; entered in MW.

ql::Vector{Float64}

Reactive power component of constant MVA load; entered in Mvar.

ip::Vector{Float64}

Active power component of constant current load; entered in MW at one per unit voltage.

iq::Vector{Float64}

Reactive power component of constant current load; entered in Mvar at one per unit voltage.

yp::Vector{Float64}

Active power component of constant admittance load; entered in MW at one per unit voltage.

yq::Vector{Float64}

Reactive power component of constant admittance load; entered in Mvar at one per unit voltage. YQ is a negative quantity for an inductive load and positive for a capacitive load.

owner::Vector{Int64}

Owner to which the load is assigned (1 through the maximum number of owners at the current size level).

PowerFlowData.Generators — Type

struct Generators <: PowerFlowData.Records

Each network bus to be represented as a generator or plant bus in PSS/E must be specified in a generator data record. In particular, each bus specified in the bus data input with a type code of two (2) or three (3) must have a generator data record entered for it.

Fields

i::Vector{Int64}

Bus number, or extended bus name enclosed in single quotes.

id::Vector{WeakRefStrings.String3}

One- or two-character uppercase non blank alphanumeric machine identifier used to distinguish among multiple machines at bus "I". It is recommended that, at buses for which a single machine is present, the machine be designated as having the machine identifier ’1’. ID = ’1’ by default.

pg::Vector{Float64}

Generator active power output; entered in MW. PG = 0.0 by default.

qg::Vector{Float64}

Generator reactive power output; entered in Mvar. QG needs to be entered only if the case, as read in, is to be treated as a solved case. QG = 0.0 by default.

qt::Vector{Float64}

Maximum generator reactive power output; entered in Mvar. For fixed output gen- erators (i.e., nonregulating), QT must be equal to the fixed Mvar output. QT = 9999.0 by default.

qb::Vector{Float64}

Minimum generator reactive power output; entered in Mvar. For fixed output generators, QB must be equal to the fixed Mvar output. QB = -9999.0 by default.

vs::Vector{Float64}

Regulated voltage setpoint; entered in pu. VS = 1.0 by default.

ireg::Vector{Int64}

Bus number, or extended bus name enclosed in single quotes, of a remote type 1 or 2 bus whose voltage is to be regulated by this plant to the value specified by VS. If bus IREG is other than a type 1 or 2 bus, bus "I" regulates its own voltage to the value specified by VS. IREG is entered as zero if the plant is to regulate its own voltage and must be zero for a type three (swing) bus. IREG = 0 by default.

mbase::Vector{Float64}

Total MVA base of the units represented by this machine; entered in MVA. This quantity is not needed in normal power flow and equivalent construction work, but is required for switching studies, fault analysis, and dynamic simulation. MBASE = system base MVA by default.

zr::Vector{Float64}

Complex machine impedance, ZSORCE; entered in pu on MBASE base. This data is not needed in normal power flow and equivalent construction work, but is required for switching studies, fault analysis, and dynamic simulation. For dynamic simulation, this impedance must be set equal to the unsaturated subtransient impedance for those generators to be modeled by subtransient level machine models, and to unsaturated transient impedance for those to be modeled by classical or transient level models. For short-circuit studies, the saturated subtransient or transient impedance should be used. ZR = 0.0 by default.

zx::Vector{Float64}

See zr. ZX = 1.0 by default.

rt::Vector{Float64}

Step-up transformer impedance, XTRAN; entered in pu on MBASE base. XTRAN should be entered as zero if the step-up transformer is explicitly modeled as a network branch and bus "I" is the terminal bus. RT+jXT = 0.0 by default.

xt::Vector{Float64}

See rt. RT+jXT = 0.0 by default.

gtap::Vector{Float64}

Step-up transformer off-nominal turns ratio; entered in pu. GTAP is used only if XTRAN is nonzero. GTAP = 1.0 by default.

stat::Vector{Bool}

Initial machine status of one for in-service and zero for out-of-service. STAT = 1 by default.

rmpct::Vector{Float64}

Percent of the total Mvar required to hold the voltage at the bus controlled by this bus "I" that are to be contributed by the generation at bus "I"; RMPCT must be positive. RMPCT is needed if IREG specifies a valid remote bus and there is more than one local or remote voltage controlling device (plant, switched shunt, FACTS device shunt element, or VSC dc line converter) controlling the voltage at bus IREG to a setpoint. RMPCT is needed also if bus "I" itself is being controlled locally or remotely by one or more other setpoint mode voltage controlling devices. RMPCT = 100.0 by default.

pt::Vector{Float64}

Maximum generator active power output; entered in MW. PT = 9999.0 by default.

pb::Vector{Float64}

Minimum generator active power output; entered in MW. PB = -9999.0 by default.

oi::Vector{Int64}

Owner number; (1 through the maximum number of owners at the current size level). Each machine may have up to four owners. By default, O1 is the owner to which bus "I" is assigned and O2, O3, and O4 are zero.

fi::Vector{Float64}

Fraction of total ownership assigned to owner Oi; each Fi must be positive. The Fi values are normalized such that they sum to 1.0 before they are placed in the working case. By default, each Fi is 1.0.

PowerFlowData.Branches — Type

struct Branches <: PowerFlowData.Records

In PSS/E, the basic transmission line model is an Equivalent Pi connected between network buses.

Data for shunt equipment units, such as reactors, which are connected to and switched with the line, are entered in the same data record.

Shunts connected to buses

To represent shunts connected to buses, that shunt data should be entered in the bus data record.

Transformers

Branches to be modeled as transformers are not specified in this data category; rather, they are specified in the transformer data category.

Fields

i::Vector{Int64}

Branch "from bus" number, or extended bus name enclosed in single quotes.

j::Vector{Int64}

Branch "to bus" number, or extended bus name enclosed in single quotes. "J" is entered as a negative number, or with a minus sign before the first character of the extended bus name, to designate it as the metered end; otherwise, bus "I" is assumed to be the metered end.

ckt::Vector{WeakRefStrings.String3}

One- or two-character uppercase nonblank alphanumeric branch circuit identifier; the first character of CKT must not be an ampersand "&". It is recommended that single circuit branches be designated as having the circuit identifier '1'. CKT = '1' by default.

r::Vector{Float64}

Branch resistance; entered in pu. A value of R must be entered for each branch.

x::Vector{Float64}

Branch reactance; entered in pu. A nonzero value of X must be entered for each branch.

b::Vector{Float64}

Total branch charging susceptance; entered in pu. B = 0.0 by default.

rate_a::Vector{Float64}

First loading rating; entered in MVA. If RATEA is set to 0.0, the default value, this branch will not be included in any examination of circuit loading.

Ratings are entered as: $MVA_{rated} = sqrt(3) × E_{base} × I_{rated} × 10^{-6}$ where:

$E_{base}$ is the base line-to-line voltage in volts of the buses to which the terminal of the branch is connected.

$I_{rated}$ is the branch rated phase current in amperes.

rate_b::Vector{Float64}

Second loading rating; entered in MVA. RATEB = 0.0 by default.

rate_c::Vector{Float64}

Third loading rating; entered in MVA. RATEC = 0.0 by default.

gi::Vector{Float64}

Complex admittance of the line shunt at the bus "I" end of the branch; entered in pu. BI is negative for a line connected reactor and positive for line connected capacitor. GI + jBI = 0.0 by default.

bi::Vector{Float64}

Complex admittance of the line shunt at the bus "I" end of the branch; entered in pu. BI is negative for a line connected reactor and positive for line connected capacitor. GI + jBI = 0.0 by default.

gj::Vector{Float64}

Complex admittance of the line shunt at the bus "J" end of the branch; entered in pu. BJ is negative for a line connected reactor and positive for line connected capacitor. GJ + jBJ = 0.0 by default.

bj::Vector{Float64}

Complex admittance of the line shunt at the bus "J" end of the branch; entered in pu. BJ is negative for a line connected reactor and positive for line connected capacitor. GJ + jBJ = 0.0 by default.

st::Vector{Bool}

Initial branch status where 1 designates in-service and 0 designates out-of-service. ST = 1 by default.

len::Vector{Float64}

Line length; entered in user-selected units. LEN = 0.0 by default.

oi::Vector{Int64}

Owner number; 1 through the maximum number of owners at the current size level. Each branch may have up to four owners. By default, O1 is the owner to which bus "I" is assigned and O2, O3, and O4 are zero.

fi::Vector{Float64}

Fraction of total ownership assigned to owner Oi; each Fi must be positive. The Fi values are normalized such that they sum to 1.0 before they are placed in the working case. By default, each Fi is 1.0.

PowerFlowData.TwoWindingTransformers — Type

struct TwoWindingTransformers <: PowerFlowData.Records

Each AC transformer to be represented in PSS/E is introduced through transformer data records that specify all the data required to model transformers in power flow calculations, with one exception.

That exception is a set of ancillary data, comprising transformer impedance correction tables, which define the manner in which transformer impedance changes as off-nominal turns ratio or phase shift angle is adjusted. Those data records are described in Transformer Impedance Correction Tables.

Currently only two-winding transformers are supported.

In PSS/E files, both two-winding and three-winding transformers are specified in transformer data records. The fields for the two-winding transformers are common to the three-winding transformers; data for two-winding transformers is a subset of the data required for three-winding transformers. Currently all data in the transformer records is assumed to be for two-winding transformers.

Fields

i::Vector{Int64}

j::Vector{Int64}

k::Vector{Bool}

ckt::Vector{WeakRefStrings.String3}

cw::Vector{Int64}

cz::Vector{Int64}

cm::Vector{Int64}

mag1::Vector{Float64}

mag2::Vector{Float64}

nmetr::Vector{Int64}

name::Vector{WeakRefStrings.String15}

stat::Vector{Bool}

oi::Vector{Int64}

fi::Vector{Float64}

r1_2::Vector{Float64}

x1_2::Vector{Float64}

sbase1_2::Vector{Float64}

windv1::Vector{Float64}

nomv1::Vector{Float64}

ang1::Vector{Float64}

rata1::Vector{Float64}

ratb1::Vector{Float64}

ratc1::Vector{Float64}

cod1::Vector{Int64}

cont1::Vector{Int64}

rma1::Vector{Float64}

rmi1::Vector{Float64}

vma1::Vector{Float64}

vmi1::Vector{Float64}

ntp1::Vector{Int64}

tab1::Vector{Int64}

cr1::Vector{Float64}

cx1::Vector{Float64}

windv2::Vector{Float64}

nomv2::Vector{Float64}

Alternatives

In Julia, I am are aware of two other open-source packages with functionality to parse PSS/E files:

PowerModels.jl

PowerSystems.jl

I have not used either so cannot recommend one over the other. From what I can see, these parsers are almost identical to each other. It seems PowerSystems.jl originally vendored the PowerModels.jl code, but the parsers may have diverged slightly over time.

Importantly, these alternatives take a completely different approach to this package. These other parsers read the .raw files as a String (e.g. using readlines), then operate on string data, and parse strings into other Julia types as necessary.

PowerFlowData.jl (this package) reads the .raw files as a bytes buffer (Vector{UInt8}), then parses the bytes directly into Julia types.

Hopefully this will be much faster and more memory efficient, but benchmarks pending. For now, this package is being developed as a fun exercise. Feel encouraged to give this package a try (and open issues!), but these alternative parsers are surely more battle-tested!

Implementation details

We use Parsers.jl to parse bytes into Julia types. Broadly speaking, we use Parsers.Options to configure the parsing based on the .raw format (e.g. , characters are delimiters), and then Parsers.xparse to actually parse the bytes between delimiters into the expected Julia types. The expected Julia type depends on the category of data we are reading at that point in the file (buses, loads, …); if the PSS/E user manual says "load records" should come after "bus records", and each load record should have 12 columns with the first column containing an integer "bus number", then we try to parse the first value in a load record as an Int, and so on.